testing balances with Whale
