+++ screens/WalletNavigator/BottomTabNavigator.tsx
@@ -2,7 +2,7 @@ import { Ionicons } from '@expo/vector-icons'
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs'
 import * as React from 'react'
 
-import { BalancesNavigator } from './screens/BalancesScreen/BalancesScreen'
+import { BalancesNavigator } from './screens/BalancesScreen/BalancesNavigator'
 import { LiquidityNavigator } from './screens/LiquidityScreen/LiquidityScreen'
 import { TransactionsNavigator } from './screens/TransactionsScreen/TransactionsScreen'
 import { SettingsNavigator } from './screens/SettingsScreen/SettingsNavigator'
