+++ screens/WalletNavigator/screens/BalancesScreen/BalancesScreen.tsx
@@ -1,48 +1,39 @@
 import * as React from 'react'
-import { Button } from 'react-native'
+import { Button, View } from 'react-native'
 import tailwind from 'tailwind-rn'
 
 import { translate } from '../../../../translations'
-import { Text, View } from '../../../../components/Themed'
 import { getTokenIcon } from '../../../../components/icons/tokens/_index'
-import { createStackNavigator } from '@react-navigation/stack'
+import { useWhaleApiClient } from "../../../../hooks/api/useWhaleApiClient";
+import { useWalletAPI } from "../../../../hooks/wallet/WalletAPI";
+import { useEffect } from "react";
+import { AddressToken } from "@defichain/whale-api-client/dist/api/address";
 
 export function BalancesScreen (): JSX.Element {
-  const [count, setCount] = React.useState(0)
+  const whaleApiClient = useWhaleApiClient()
+  const account = useWalletAPI().getWallet().get(0)
+
+  const [utxoBalance, setUtxoBalance] = React.useState<string>("")
+  const [tokens, setTokens] = React.useState<AddressToken[]>([])
+
+  useEffect(() => {
+    account.getAddress().then(async address => {
+      await whaleApiClient.address.getBalance(address).then(value => {
+        setUtxoBalance(value)
+      })
+
+      await whaleApiClient.address.listToken(address, 30).then(tokens => {
+        setTokens(tokens)
+      })
+    })
+  }, []);
 
   const Icon = getTokenIcon('DFA')
 
   return (
     <View style={tailwind('flex-1 items-center justify-center')}>
       <Icon />
-      <Button testID='count_btn' title='Click' onPress={() => setCount(count + 2)} />
-      <Text testID='count_text'>
-        Count: {count}
-      </Text>
-      <Text style={tailwind('text-xl font-bold')}>
-        {translate('screens/BalancesScreen', 'Balances')}
-      </Text>
-      <View style={tailwind('w-4/5 h-px my-8')} lightColor='#eee' darkColor='rgba(255,255,255,0.1)' />
+
     </View>
-  )
-}
-
-export interface BalancesParamList {
-  BalancesScreen: undefined
-
-  [key: string]: undefined | object
-}
-
-const Balances = createStackNavigator<BalancesParamList>()
-
-export function BalancesNavigator (): JSX.Element {
-  return (
-    <Balances.Navigator>
-      <Balances.Screen
-        name='balances'
-        component={BalancesScreen}
-        options={{ headerTitle: translate('screens/BalancesScreen', 'Balances') }}
-      />
-    </Balances.Navigator>
   )
 }
